requirements.txt
discord.py==2.3.2
mysql-connector-python
GitPython
requests